+++ NEWS.md
@@ -5,6 +5,11 @@
 * `seq_linter()` additionally lints on `1:n()` (from dplyr) 
   and `1:.N` (from data.table)  (#1396, @IndrajeetPatil).
 
+## New features
+
+* New `function_argument_linter()` that checks that arguments with defaults appear last in function declarations 
+  (#450, @AshesITR).
+
 ## Bug fixes
 
 * `get_source_expressions()` no longer fails on R files that match a knitr pattern (#743, #879, #1406, @AshesITR).